+++ src/main/java/net/neganote/gtutilities/common/item/OmniBreakerItem.java
@@ -84,6 +84,7 @@ public class OmniBreakerItem extends ComponentItem {
         tooltipComponents.add(Component.translatable("tooltip.omnibreaker.charge_status",
                 Component.translatable(FormattingUtil.formatNumbers(charge)).withStyle(ChatFormatting.GREEN),
                 Component.translatable(FormattingUtil.formatNumbers(maxCharge)).withStyle(ChatFormatting.YELLOW)));
+        tooltipComponents.add(Component.translatable("tooltip.omnibreaker.modern_vajra"));
         tooltipComponents
                 .add(Component.translatable("tooltip.omnibreaker.can_break_anything").withStyle(ChatFormatting.GRAY));
         tooltipComponents