+++ changelog.md
@@ -1,3 +1,5 @@
-# GregTech Modern Utilities version 1.1.3
+# GregTech Modern Utilities version 1.1.4
 
-Fix Omni-breaker not looking right in third person
+* Fix PTERB coolant not being re-enabled if only one of the coolant vars was zero
+* Add more Jade integration, putting the PTERB's coolant drain in the hover info
+* Add tool tags to Omni-breaker, update tooltip